requirements.txt
Flask==2.1.1
Flask-PyMongo==2.3.0
Flask-Mail==0.9.1
python-dotenv==0.19.2
itsdangerous==2.0.1
Flask-Mail
itsdangerous